+++ packages/veui/src/components/Dialog.vue
@@ -3,10 +3,12 @@
   ref="overlay"
   class="veui-dialog"
   :open="localOpen"
-  :overlay-class="mergeOverlayClass({
-    'veui-dialog-box': true,
-    'veui-dialog-box-mask': modal
-  })"
+  :overlay-class="
+    mergeOverlayClass({
+      'veui-dialog-box': true,
+      'veui-dialog-box-mask': modal
+    })
+  "
   :ui="realUi"
   autofocus
   :modal="modal"
@@ -22,8 +24,12 @@
     @keydown.esc="handleEscape"
   >
     <div
-      v-if="title || $slots.title"
-      v-drag:content.translate="{ draggable, containment: '@window', ready: dragReady }"
+      v-if="title || $slots.title || $scopedSlots.title"
+      v-drag:content.translate="{
+        draggable,
+        containment: '@window',
+        ready: dragReady
+      }"
       class="veui-dialog-content-head"
       :class="{ 'veui-dialog-draggable': draggable }"
     >
@@ -57,14 +63,14 @@
           :ui="uiParts.ok"
           @click="close('ok')"
         >
-          {{ t('ok') }}
+          {{ t("ok") }}
         </veui-button>
         <veui-button
           :ui="uiParts.cancel"
           autofocus
           @click="cancel"
         >
-          {{ t('cancel') }}
+          {{ t("cancel") }}
         </veui-button>
       </slot>
     </div>
@@ -145,15 +151,15 @@ export default {
     }
   },
   methods: {
-    dragReady (handle) {
-      this.dragHandle = handle
+    dragReady ({ reset }) {
+      this.resetDrag = reset
     },
     resetPosition () {
-      if (!this.dragHandle) {
+      if (!this.resetDrag) {
         throw new Error('The dialog is not ready for drag.')
       }
 
-      this.dragHandle.reset()
+      this.resetDrag()
     },
     focus () {
       let { overlay } = this.$refs