+++ packages/veui/test/utils.js
@@ -1,3 +1,5 @@
+import { trim } from 'lodash'
+
 export function wait (timeout) {
   return new Promise(resolve => {
     setTimeout(resolve, timeout)
@@ -13,4 +15,25 @@ export function nextFrame () {
 export async function doubleNextFrame () {
   await nextFrame()
   await nextFrame()
+}
+
+export function getStyle (element) {
+  let cssText = element.style.cssText || ''
+  let arr = cssText.split(';')
+  return arr.reduce((obj, item) => {
+    if (!item) {
+      return obj
+    }
+    let [key, value] = item.split(':')
+    return { ...obj, [trim(key)]: trim(value) }
+  }, {})
+}
+
+export function normalizeTransform (transform) {
+  let el = document.createElement('div')
+  el.style.transform = transform
+  document.body.appendChild(el)
+  let result = getComputedStyle(el).transform
+  el.parentElement.removeChild(el)
+  return result === 'none' ? 'matrix(1, 0, 0, 1, 0, 0)' : result
 }